refactor(token): 함수 분리
