+++ parse/is_valid_syntax.c
@@ -3,7 +3,7 @@
 bool	is_pair_of_parenthesis(t_token *token, char **token_value)
 {
 	t_token	*temp;
-	int 	parenthesis_count;
+	int		parenthesis_count;
 
 	parenthesis_count = 0;
 	temp = token;
@@ -43,40 +43,11 @@ bool	is_valid_redirection(t_token *token, char **token_value)
 	return (true);
 }
 
-t_token	*get_last_token_in_parenthesis(t_token **token)
-{
-	t_token	*temp;
-	int		parenthesis_count;
-
-	if (!token || !*token)
-		return (NULL);
-	parenthesis_count = 1;
-	temp = *token;
-	while (temp->next)
-	{
-		if (temp->next->type == PARENTHESIS_OPEN)
-			parenthesis_count++;
-		else if (temp->next->type == PARENTHESIS_CLOSE)
-			parenthesis_count--;
-		if (parenthesis_count == 0)
-			break ;
-		temp = temp->next;
-	}
-	if (!temp->next)
-		return (NULL);
-	if (temp->type == PARENTHESIS_OPEN || temp->type == PARENTHESIS_CLOSE)
-	{
-		*token = temp;
-		return (NULL);
-	}
-	return (temp);
-}
-
 bool	is_valid_parenthesis(t_token *token, char **token_value)
 {
 	t_token	*prev;
-	t_token *next;
-	t_token *temp;
+	t_token	*next;
+	t_token	*temp;
 
 	prev = NULL;
 	temp = token;
@@ -103,7 +74,7 @@ bool	is_valid_parenthesis(t_token *token, char **token_value)
 bool	is_valid_command(t_token *token, char **token_value)
 {
 	t_token	*temp;
-	bool 	is_in_normal;
+	bool	is_in_normal;
 
 	temp = token;
 	is_in_normal = false;
